[CPDLP-2961] Group Application migrators
