+++ app/services/migration/migrator.rb
@@ -13,7 +13,6 @@ module Migration
       Migration::Migrators::School,
       Migration::Migrators::Course,
       Migration::Migrators::Application,
-      Migration::Migrators::ApplicationNotInEcf,
     ].freeze
 
     class << self
